Método weight adicionado na interface e outras refatorações.
* Representação do infinito no lugar de Float.MAX_VALUE
* Método GraphStorage#normalize removido

+++ A1/src/test/java/br/ufsc/graphs/structures/WeightedGraphTest.java
@@ -30,7 +30,7 @@ class WeightedGraphTest {
 
     @Test
     void weight() {
-        assertEquals((float)0.3, graph.weight(0, 2));
+        assertEquals(0.3, graph.weight(0, 2));
     }
 
     @Test
